requirements.txt
streamlit
numpy
pillow
matplotlib
scikit-learn
tensorflow==2.15.0